don't inline functions if patching is enabled
Summary:
Context: In Strict Python, we keep track of whether patching is enabled via the `enable_patching` flag to support testing workflows. However, inlining in static python can break this behavior. Previously, we had gotten around this by not allowing patching to be enabled if sys.optimize is set (this is the only time where Static Python decides to inline at the moment). However, there are some cases where we'd like to allow patching even if sys.optimize is equal to 2. To support these cases, thread Strict Python's patching enabled flag to the Static Python compiler, and refuse to inline.

One obvious issue this causes is creating different bytecode based on whether patching is enabled. To ensure that we don't cache incorrect bytecode, introduce a separate prefix to the strict loader when patching is enabled, but keep the old .pyc's the same.

Reviewed By: carljm

Differential Revision: D30965210

fbshipit-source-id: 71760d22a1