never insert casts to object for return types
Summary:
Ensure that we never insert useless casts to `object` when returning a dynamic
value from a function which is annotated to return dynamic (i.e. unknown) type,
annotated to return `object`, or not annotated with a return type at all, whether
the function is sync or async.

Previously async was broken because we checked for dynamic expected return
before unwrapping `Awaitable`. And annotated `object` always inserted a cast.

This also required fixing the insertion of `Awaitable` return type for async functions
so that in the case of a function with no annotated return type, we still mark the
return type as `Awaitable[DYNAMIC]` instead of just `DYNAMIC`. The no-annotation
case should be consistent with the unknown-annotation case, and in both cases
since the function is async we do know it returns an awaitable.

Reviewed By: wmeehan

Differential Revision: D31175817

fbshipit-source-id: d124c7c588