check incompatible field overrides
Summary:
We weren't checking for incompatible field overrides at all. We were still sound,
because we actually just ignored field overrides entirely and went with the base
field type, but it's an odd user experience.

In principle we could allow type narrowing of attributes on subclasses, only if the
attribute is `Final`. But I think this would require changes in the runtime.

Reviewed By: DinoV

Differential Revision: D31177531

fbshipit-source-id: 7cd9742240