introduce module types to our type system
Summary:
Context: Static Python does a good job of optimizing uses of values when they're imported via the form `from a.b import value`, however it has historically missed accesses via `a.b.value`. This is because in order to properly analyze `a.b.value`, we need to know that `a` and `a.b` are modules. We were previously not including these in the type system, which meant that we would lose all precision on these values.

This diff introduces module types and starts using them to resolve binding attrs - the places where we need to take most care is ensuring that we get the imports of `import a.b` correct, as well as handling the `as` aliasing syntax well.

Reviewed By: carljm

Differential Revision: D31149959

fbshipit-source-id: 6b256975ac