only resolve base static functions when checking return type
Summary: This change addresses a bug where a chain of non-static classes inheriting from a static class have two non-static overrides of the static base. In that case, we'd lose the expected static return type, and crash. To fix this, let's only look at static bases in the MRO when getting derived functions with the intent of resolving the return type.

Reviewed By: DinoV

Differential Revision: D31180200

fbshipit-source-id: f6f80e4f2e