add method for testing linting
Summary:
Ran into a bug where we were failing to set a type on a node in case we are
lint-typechecking (i.e. collecting errors with `CollectingErrorSink`), so added
a utility for "linting" a code string, collecting the errors from it, and matching
the errors to locations with `at` sugar.

Reviewed By: sinancepel

Differential Revision: D31227825

fbshipit-source-id: 2d401f3e75