use consistent pyc file name
Summary:
In D30965210 (https://github.com/facebookincubator/cinder/commit/a795d531ea9f9ce04ef82cd679c8f7217f523eb4) we forgot to pass in `self.enable_patching` in two out of three
call-sites, so we weren't writing to the same pyc file name as we were reading
from (when patching was enabled.) This was half of the cause of S245134 which
required a Cinder version revert in D31250447. This fixes that, and D31252873
fixes the other half of the problem.

Also make `enable_patching` a required argument to `add_strict_tag` so we can't
forget it at any other call site.

Differential Revision: D31248255

fbshipit-source-id: 6839e692ff