Add py-walk-threads to cinder gdb extension
Summary:
Add `py-walk-threads`, which loops through threads, running `py-walk-frame` on each.
This is very similar to `t a a py-walk-frame`, but the output is cleaner when you have a lot of threads where the majority are not python.

It's easier to explain to those unfamiliar with gdb, and easier to call via automated tools.

While I was here I also included a bug fix for `t a a py-walk-frame`, where it would stop after the first thread that didn't contain python frames as raising a gdb exception stops thread iteration.

Reviewed By: czardoz

Differential Revision: D31256763

fbshipit-source-id: e0b62352d9