Fix crash in multiple inheritance scenarios where v-table isn't initialized
Summary:
We have an issue where we can fail to initialize a v-table in face of multiple inheritance coming from non-static classes.

What happens is that when we go to recursively init the v-tables for subclasses via `type_init_subclass_vtables` we call to initialize a class via `_PyClassLoader_EnsureVtable` and the immediate base class for that is not initialized, so we recursively initialize it.

But that base class has other subclasses which don't get initialized, so when we continue recursing through the subclass hierarchy in `type_init_subclass_vtables` we end up skipping it because it was already initialized.

We then crash in `_PyClassLoader_UpdateDerivedSlot` when the missing v-table isn't there.

Now `_PyClassLoader_EnsureVtable` is aware if we're initializing subclasses, and if causes the base class to be initialized, we'll go ahead and initialize the subclasses for the base class as well.

Reviewed By: carljm

Differential Revision: D31279020

fbshipit-source-id: ec00db2f2f