Set FunctionDef node type to Function or DecoratedMethod
Summary:
This changes the type we set for the `FunctionDef` / `AsyncFunctionDef` nodes to be a `FunctionContainer` object rather than being the return type.  We then use the type of the node to generate the function.

`Function` objects themselves are `FunctionContainer` objects, and our `DecoratedMethod` objects are also `FunctionContainer`s.  This lets each decorated method control its own binding and code generation.  They also get a `bind_function_self` which lets the handling of the argument binding for `self` be handled by the decorator instance as well.

After this we no longer need the `dynamic_returns` because we already set the `Function` objects return value.

Reviewed By: sinancepel

Differential Revision: D30848746

fbshipit-source-id: 3778628a8f