Support ContextDecorator stacked with other built-in decorators
Summary:
This lets us stack our decorators with the `ContextDecorator`.  The key thing here is now that when we apply known decorators is that each decorator object is responsible for emitting the decorator expression, calling the inner decorator to emit the function, and then calling the decorator on the resulting function.  Thus for known decorators we never hit the loop over the decorators on a function and instead apply them recursively.

So decorators like `cached_property` where the decorator is not applied at all until after class creation time can simply skip generating the decorator emission and call.

It also modifies the check for if a decorator is applicable - now instead of looking at internal implementation type we look at the type of the object within the Python type system to see that it's a function. This lets us have the resulting type of a context decorator still be a Python function and thus it is now consumable by the decorators as well.

We also need to deal with `staticmethod` and `classmethod` which are going to alter the parameter of the 1st argument.  They're a little bit odd in that they're not really changing the function, but they are changing how we interpret the function.   For this there's a new `replace_function` which lets us push down the updates to the inner decorators.  This also means that we're now properly capturing the type of the 1st argument during the declaration visit for these methods, which we weren't before.  Ultimately we need to switch to using the types discovered during the declaration visit as the truth on the function objects instead of re-visiting it in the `TypeBinder`, but that's a fix for another day!

Reviewed By: carljm

Differential Revision: D30965674

fbshipit-source-id: 7a5f0e047b