default nodes to DYNAMIC in lint mode
Summary:
When we have a non-throwing error sink (lint mode) to collect all errors, previously
we had to carefully ensure that in all error-throwing cases we still fell back to setting
the node type to DYNAMIC, so we wouldn't later run afoul of the check in `get_type`
that all nodes have a type.

Instead, default node type to DYNAMIC when in lint mode. Only in the type binder,
not in codegen, since we should never codegen after a lint-mode type binding, and
in codegen we want to continue to ensure all nodes were given an explicit type.

Removed one particularly awkward DYNAMIC-fallback to verify this works; there are
probably other cases where we now needlessly fall back to DYNAMIC on an error,
but there's no particular need to remove them all.

Reviewed By: DinoV

Differential Revision: D31348980

fbshipit-source-id: cf4bc9da6d