Ensure dynamic_return works correctly
Summary:
The `dynamic_return` decorator is currently broke, because it doesn't
return a TransparentDecoratorMethod. Additionally, it doesn't account
for the fact that the inferred return type for coroutine functions
should be an `AwaitableType`. This diff fixes both issues (and adds a
testcase for them).

Also, moved the dynamic return tests into their own module.

Reviewed By: carljm

Differential Revision: D31357528

fbshipit-source-id: 7f618fddf4