add revealed_type() utility
Summary:
When we have tests that are just asserting on the bound type of something,
it's way more explicit and less error-prone to add a `reveal_type` and explicitly
assert on the type of that expression, rather than try to hack around it with
e.g. return type annotations. So let's make the better pattern a bit easier
with a dedicated utility. Converted the union tests to use it.

Reviewed By: DinoV

Differential Revision: D31451894

fbshipit-source-id: 88956f91f4