rename SymbolTable to Compiler
Summary:
The `SymbolTable` class long ago outgrew simply being a symbol table; it manages
the entire process of Static Python compilation. Give it the more appropriate title of
`Compiler`.

For now this makes the naming a bit awkward in `strict/compiler.py` because we have
a clash between strict modules' idea of a `Compiler` and Static Python's. I have ideas
to resolve this better but will do it in a separate diff to avoid renaming too many things
at once.

Reviewed By: DinoV

Differential Revision: D31454446

fbshipit-source-id: e317e33b97