simplify strict.Compiler to always support static python
Summary:
We always use the static-supporting compiler in our import loader; we don't expose any
way to use strict modules without also supporting the possibility of static python.
So we can merge `Compiler` and `StaticCompiler` into a single `Compiler` class that
always supports strict modules and static python. This doesn't fully resolve the
naming conflict between `strict.Compiler` and `static.Compiler`, but it simplifies the
naming in the `strict.compiler` module a lot; now we just have the strict `Compiler` and
the static one, imported as `StaticCompiler`.

Reviewed By: DinoV

Differential Revision: D31457879

fbshipit-source-id: f8e62f02fb