fix multithreaded_test_compile so most tests pass
Summary:
This has been broken for a long time, and broke even worse last week, but it's a silent breakage;
the job seems to not report failure if the tests fail, only if there's a TSAN failure. But if the tests
aren't running to completion then we'll probably miss lots of TSAN failures, so having them broken
is obviously not good.

Reviewed By: tekknolagi

Differential Revision: D31483768

fbshipit-source-id: 6660743f05