add optimize to import_module() signature
Summary:
In the Python import system, the value for `optimize` is provided per-import, not as a global.

But previous to this diff, the `import_module` API on a static Compiler did not accept a value
for `optimize`. This forced the implementation of `StrictStaticCompiler` to do something ugly:
it had an instance attribute `optimize` initialized to `None` and then updated from outside the
class whenever an import happened.

Clean this up by adding `optimize` to the `import_module` signature. This also requires threading
`optimize` through `DeclarationVisitor`, since that's where we trigger recursive analysis of
imported modules and call `import_module`.

Also remove default value of `0` for `optimize` in all the core compiler APIs, in order to avoid bugs
where we accidentally don't pass on the correct `optimize` value and default it back to 0. This
makes use of some of these methods (e.g. `compile`, `bind`, `add_module`) more verbose from
tests, but I plan to ensure we have test-specific utility wrappers to make these patterns in tests
less verbose anyway, so we shouldn't need to directly call `Compiler` methods from most
individual tests.

Reviewed By: DinoV

Differential Revision: D31461159

fbshipit-source-id: abd7be6745