remove strict_features_enabled flag effects
Summary:
The `strict_features_enabled` flag was a temporary shim to allow smooth migration of IGSRV
from the IGSRV-side loader and strict modules to the Cinder-side ones. This migration has
now happened, so we can default the flag to `True`, or in fact just remove all the checks of
the flag and assume it.

It would probably work fine to remove the definition of the flag also (IGSRV would just set an
otherwise nonexistent attribute), but I think it's clearer if we let it continue to exist until IGSRV
stops setting it (which can happen after we ship a release to IGSRV including this diff.)

Reviewed By: czardoz

Differential Revision: D31462148

fbshipit-source-id: 5e753bd25c