use more of Compiler in bind_module
Summary:
Call `Compiler._bind` instead of duplicating it in `StaticTestBase.bind_module`.

This means we no longer have access to a `TypeBinder` to return to the caller,
but only one test was using this, and it was too deep into implementation
details; rewrote that test to just verify externally that correct location info is
attached to errors from within inline funcs.

It also means that `bind_module` now matches real compilation in terms of
optimizer usage, meaning that `test_type_binder` needed some updates to
mark things as actually generating literals due to constant folding.

Reviewed By: czardoz

Differential Revision: D31463042

fbshipit-source-id: 1e6bb650e1