reuse Compiler methods in StaticCodeGenerator.make_code_gen
Summary:
Instead of having `StaticCodeGenerator.make_code_gen` duplicate the functionality
of `Compiler`, have it use a `Compiler` instead.

This means we can't support static compilation without the AST optimizer or the peephole
optimizer, but that's fine, those options only exist for testing the AST optimizer and
peepholer, we don't use or need those options for Static Python compilation.

This means we always compile with knowledge of built-in modules, which can cause
us to fail to decl-visit if we try to compile something with the same modname as a
built-in module.

Reviewed By: czardoz

Differential Revision: D31463512

fbshipit-source-id: 8623c0b05b