Rename `TypeProfiler` to `FixedTypeProfiler`, add new `TypeProfiler`
Summary:
Except for the name change, `FixedTypeProfiler` is the same as before:
a fixed number of slots with one type per slot.

`TypeProfiler` is more dynamic and flexible: it's a 2-dimensional type profiler
with the number of rows and columns determined at creation time. It's intended
to be used in situations where a vector of types should be profiled together
(i.e., the inputs to an opcode like `BINARY_ADD`).

Reviewed By: mpage

Differential Revision: D31067400

fbshipit-source-id: cc49ef1a70