don't add v-table entries for non-static types
Summary:
Currently we'll initialize slots for static methods that  show up in non-static types.  One source of this could be a decorator that's defined in a static module, that then gets applied in a non-static module, and returns a wrapper method.

if the type then gets modified we'll try and find the associated static function which we're overriding, which doesn't exist because we're a non-static type.  In which case we crash.

This causes us not to fill in slots on non-static types - they can't possibly offer any new overrides.  The check for a non-static type is a little bit difficult because we want to include built-in types and our generic instantiations.  We also have a couple of types which are basically static which aren't currently marked as such, so this updates those as well.

Reviewed By: czardoz

Differential Revision: D31476336

fbshipit-source-id: d1103a94d7