Add resolve_descr_get API
Summary:
This adds a `resolve_descr_get` API which doesn't take a TypeBinder and returns the actual value instead of publishing it on the TypeBinder.  There's a base `bind_descr_get` API which defaults to using this, but classes can still override `bind_descr_get` if they need the `TypeBinder`.

This is the start of building up a small set of `resolve_*` APIs which can be shared between the `TypeBinder` and `DeclarationVisitor`

Reviewed By: carljm

Differential Revision: D31580711

fbshipit-source-id: aca8053481