run in_strict_module tests through strict rewrite
Summary:
This runs all `in_strict_module` Static Python tests through the strict modules rewrite,
to bring them closer to prod behavior. We've had a number of bugs not caught in our
test suite because we didn't integrate strict modules rewrite with static compile until
release.

We use a fake `StrictModulesLoader` to avoid running the test code through the actual
strict analysis. This keeps the tests faster, avoids needing a pile of strict module stub
files in our test suite (for `__static__` and all its dependencies), and circumvents the
fact that strict analysis leaks memory in reference cycles, which makes the refleak
tests unhappy if used in Python test suite.

Skipping analysis (and thus AST preprocessing) will cause some differences in the
rewriter, but only for edge cases like cached properties. The main effect of the rewriter
(fixed modules and builtins) is still the same as in prod.

If in future we follow through on eliminating the strict rewriter and moving all its
features into the compiler, then we probably can undo this change.

Differential Revision: D31489017

fbshipit-source-id: d30a8450fb