add in_module() to TestCompiler
Summary:
Add an `in_module()` context manager method to `TestCompiler`, to allow for cross-module tests
that actually import the test modules. After constructing a compiler, you give a single module
name to `in_module()`. The modules you provided to the compiler will be compiled and imported
in the order you provided them, until the requested module is reached, and it is yielded.

This means you must provide the modules to the compiler in dependency order; that is, modules
that will be imported before the modules that import them.

Perhaps ideally we would instead do an import visit of all provided module sources to construct a
module dependency graph, and then topo sort and automatically import only the needed modules
in the correct order. Or even better, temporarily install an import loader that knows how to import
“modules” from the test compiler.  But this seems like too much trouble for a test utility :) In practice it is intuitive
to just provide the modules in the correct order.

This diff does go to a bit of trouble to cache optimized ASTs for the given module sources, so that
if you do provide modules in the wrong order, you'll get a (somewhat) clearer "no module named x"
error on import, rather than a mystifying missing node type error in binding (due to different ASTs
being used for the same module, one on `add_module` when it is decl-visited as a compilation
dependency and then a different one when it is actually compiled.)

Reviewed By: sinancepel

Differential Revision: D31688834

fbshipit-source-id: ca890ff071