refactor Class.bind_call
Summary:
In working on a later diff, I had some modifications to this function, and Pyre
complained that it was too complex for type inference. I think this is mostly
due to difficulty with loops. I resolved it by extracting a `_resolve_dunder`
method. Applying that refactor separately here.

One side effect of the assert in `_resolve_dunder` is that now we enforce that
MROs should always end with `OBJECT_TYPE`, so I also fix a couple types
here where they didn't. (This doesn't apply to `CType` since it overrides
`bind_class` anyway.)

Reviewed By: sinancepel

Differential Revision: D31720109

fbshipit-source-id: 795ad87a6d