emit INVOKE_FUNCTION with correct namespace
Summary:
In a scenario like the included test, where we have module `b` with `from a import f`
and then an `f()` call, we previously emitted an `INVOKE_FUNCTION ("a", "f")` -- invoking
the target function in its module of definition. At runtime or JIT compilation the classloader
will go look for `f` in the namespace of module `a`.

This works fine in the absence of patching, but it breaks Python namespace semantics
for patching. Someone may try to patch the name `f` in the namespace of `b`, and this should
work, but with the above form of invoke it doesn't work; we don't pay any attention to `b.f`
at runtime, only to `a.f`. Instead we should be emitting `INVOKE_FUNCTION ("b", "f")` so
that if `b.f` is patched our patching indirections notice it and call the patched function instead.
This invoke will work just as well in the unpatched case; the classloader will go find `b.f` which
will be the same object as `a.f`.

The difficulty in fixing this in our compiler is that we want `a.f` and `b.f` to continue to be the
same `Callable` object in our type environment, or else various things will get complicated.
So we have to reference some context outside of that `Callable` object to decide how to
emit an invoke to it.

For every `Name` and `Attribute` node, we store as extra node data the type descr for reaching
the correct namespace and name this value should be pulled from at runtime. For a `Name` node
that represents a module global it is always the current module; for an `Attribute` node, when the base
value is a module, it's the value's name in that module.

When that node is a callable type and we invoke it, we use that type descr instead of the definition
type descr of the callable.

This isn't a comprehensive fix for every opcode that might have this patching namespace
problem; it's targeted only to `INVOKE_FUNCTION` for now. But it should be extensible to
cover other cases as well (e.g. `TP_ALLOC` and `INVOKE_METHOD`.)

Reviewed By: sinancepel

Differential Revision: D31688833

fbshipit-source-id: 7f75cdf9ef