support invoking class methods from cls via `INVOKE_METHOD`
Summary:
This diff adds support for generalizing classmethod INVOKE support to support invoking from both classes and instances, and appropriately calls `INVOKE_FUNCTION` or `INVOKE_METHOD` in all cases.

The general idea is that we change the design of invoking a classmethod via `INVOKE_METHOD` to expect a type to be present as the `self` argument, instead of a self instance. This violates pre-existing assumptions in the code, and the consequences are:
1. In order to support replacing instances with types, we introduce a new `LOAD_TYPE` opcode that replaces the top of the stack with type(top of stack).
2. Now that you won't always get `self` as an instance, we pack information about whether a invoke has self-as-type in `nargsf`. This allows the JIT and interpreter to get the vtable from the object directly, instead of its type, where appropriate.
3. We change the type descriptors of classmethods to a special form, in order to help the JIT & interpreter know when `self` is an instance vs. a type.
4. We bump the importlib magic number as INVOKE_METHOD with classmethods has a new meaning now.

Reviewed By: DinoV

Differential Revision: D31251161

fbshipit-source-id: 4a9dd7045f