Add interpreter type profiling
Summary:
This diff adds interpreter-based type profiling to Cinder. It can be
enabled and disabled at runtime, and should have no overhead when disabled
(since it's hidden behind an existing flag we're already checking). Since it is
in the interpreter, Cinder should be started with the JIT disabled when it will
be used.

This subsumes and replaces the old interpreter cost profiling. The same
information is available in a slightly different form; as far as I know there's
no existing tooling that needs updating. Similar to the old cost profiling,
most of the code for this lives in `Jit/`, even though it's specifically
designed to be used with the JIT off. If we end up with much more code like
this, it may make sense to start another directory of C++ code that isn't so
JIT-specific.

High-level documentation of how to use the profiling is in `Jit/profiling.md`.

Most of the new code is in `Jit/` but this does touch a few core runtime
files. We should be able to remove many of those changes if we move to
jit-based profiling in the future.

Reviewed By: mpage

Differential Revision: D31065920

fbshipit-source-id: 1286eebe1d