fix patching of classmethods
Summary:
The basic problem:

When we invoke classmethods, we pass in the class to the INVOKE, which diverges from the usual bound method structure. For regular classmethod definitions and inheritance, we handle the vtable fine, but the problem happens with patching. When patching a classmethod, users need to provide a function that doesn't take the `cls` so that things work consistently between static and non-static python.

To fix this problem, let's change classmethod patching semantics so that the thunk_vectorcall recognized when a classmethod is being patched, and removes the first arg when threading the call in the entry point. This'll implicitly force people to patch with functions not accepting the cls.

Reviewed By: carljm

Differential Revision: D31838235

fbshipit-source-id: e45f75d5fb