Add annotation visitor
Summary:
This continues the `resolve_*` support and adds a new AnnotationVisitor which lives off of the `ModuleTable` so it can easily be fetched.  Like the `resolve_*` APIs the `visit*` APIs simply return the value which is resolved, or `None`.   To support that our `GenericVisitor` becomes `Generic` so we can specify the return type of visit (of course, nothing can see through to the return types of the dynamically dispatched `visit*` methods).

We also get `resolve_subscr` here so that we can drop the duplicated subscript code.

Reviewed By: carljm

Differential Revision: D31580706

fbshipit-source-id: ed421c4847