Fix multiple inheritance of built-in/static types
Summary:
Currently we are crashing on invokes against `StringEnum` when
it is treated as a string.  That's because we pick `Enum`'s methods
over `str`'s methods when building the v-table.  But `Enum` is not
a static type, and we'll never generate invokes against it.  So
we should go to the next static base in the MRO rather than the
immediately following class.

Reviewed By: carljm

Differential Revision: D31866460

fbshipit-source-id: 51af6151a2