revert optimized NodeVisitor
Summary:
Revert D22983160, which implemented an optimized AST visitor.

This was important for strict modules analyzer perf, but strict modules is now in C++
and doesn't need this. And it's causing compat issues (no `__dict__` on AST nodes),
and will be one more change to carry across upgrades. So let's just revert back to
upstream here.

Reviewed By: vladima

Differential Revision: D31907326

fbshipit-source-id: 68c968a96b