Add SimplifyIsTruthy and SimplifyCondBranch
Summary:
Add `SimplifyIsTruthy` to turn into a `LoadConst(0)` when its operand
is of a known `False` type.

Add `SimplifyCondBranch` to turn into a `Branch` when its operand is a
`LoadConst` with an `intSpec`. This requires us to clean up any `Phi`s whose `args`
become invalidated by the sudden loss of a predecessor block.

Put together, these new simplify passes should allow us to turn `if not
x` into a simple `Branch` when `x` is of a known "falsey" type

Reviewed By: swtaarrs

Differential Revision: D31445689

fbshipit-source-id: dbc177dbe2