Re-init vtable if already inited when type is marked as static
Summary:
If we have a subclass which gets defined after the base class gets inited, we're currently not properly initializing the v-table.  This is due to the fix in D31476336 (https://github.com/facebookincubator/cinder/commit/308e4e435d032f58618bd762d45222c6bbb5a0b0).

The issue here is that when a subtype is defined, and it's base class already has an initialized v-table, we immediately initialize the subclasses v-table.

But we don't actually set the bit that it's statically defined until after the class is created, and the v-table is initialized.

So now in the rare cases where this happens we simply clear and re-initialize the v-table when the type is statically declared.  This is one more reason why long-term we should have a different code path for creating statically defined types than the normal `LOAD_BUILD_CLASS` / `CALL_BUILD_CLASS`

Reviewed By: carljm

Differential Revision: D31969968

fbshipit-source-id: f73a9da10e