Fix descriptor bug in AttributeCache
Summary:
When the attribute is a descriptor, we have to pay more attention to
the descriptor's type in case it stops being a data descriptor, either through
an attribute assignment on its type (handled by watching the type in
`AttributeCache::fill()`) or an assignment to `__class__`. The latter case is
handled by moving the call of `jit::notifyICsTypeChanged()` from
`_PyJIT_TypeModified()` to `Runtime::notifyTypeModified()`, so it triggers in
more cases (type modification, `__class__` assignment, and type destruction).

I also changed `AttributeCache::typeChanged()` to be more conservative and
always clear everything, rather than adding the bookkeeping or other logic
necessary to detect this case and only clear what's strictly necessary.
Hopefully nobody's mutating types in a hot path anywhere.

Reviewed By: mpage

Differential Revision: D31979189

fbshipit-source-id: 484cb0f596720e7322358df95cf5bf419db647c9