Allow caching for data descriptors of non-heaptypes in JIT {load,store}attr caches
Summary:
This adds a fast path for attributes with a data descriptor that aren't
defined on a heap type. We don't actually require that the type is a heap
type if the attribute is a data descriptor; the descriptor will always
take precedence. One common case of this are `cdef` types defined
in Cython.

Reviewed By: vladima

Differential Revision: D29280350

fbshipit-source-id: 8df2e1c67e