fix patching of classmethods with INVOKE_METHOD
Summary: We weren't properly handling the INVOKE_METHOD case of patching class methods. If a class method is patched, you'll no longer have a classmethod there, but instead might be dealing with any Python callable. If that happens, we need to make sure that we type check the result & match the actual patched value by ignoring the self arg.

Reviewed By: wmeehan

Differential Revision: D32024198

fbshipit-source-id: a65a5fab04