Add @generated declaration to Jit/hir/type_generated.h
Summary:
This declaration tells Phabricator to keep the file closed by default,
preventing changes to generated files from cluttering up code review.
It also warns people not to make changes in VS Code.

The declaration must be split up in the generator so that Phabricator
does not think that it is generated as well.

Reviewed By: swtaarrs

Differential Revision: D32030044

fbshipit-source-id: 38503d26e4