Only fill ICs when the type has a valid version tag
Summary:
swtaarrs's comments about `PyType_Modified` made me nervous so I went
and took a look at our attribute-related caching code. We rely on
`PyType_Modified` to correctly invalidate caches when types are modified.
However, `PyType_Modified`, will not notify us if the type does not have
a valid version tag.

The odds of this happening are very small. Version tags are assigned
in the `_PyType_Lookup` calls that happen in the slow paths that fill the
caches. There are two possible scenarios where we could end up with a
type with without a valid version tag when filling the cache:

1. Tag assignment could fail. Tag assignment only fails either on the very
   first tag assignment at runtime startup or when the tag counter wraps
   around. The tag counter is an unsigned 32 bit value; I don't know how
   likely it is to overflow.
2. The type is modified (and no new version tag is assigned) between when
   the version tag is assigned and when we fill the cache. That could happen
   if an instance dictionary lookup executed arbitrary code.

We now only fill the cache if the type has a valid version tag.

Reviewed By: swtaarrs

Differential Revision: D32050168

fbshipit-source-id: cccab1b1d3