Add raising capability on immutability violations
Summary:
Add a way to report violations about immutability. This is similar to PyErr_RaiseCinderWarnings but I want a separate handler, and the logging will be only enabled on a small set of hosts.

The main way to use this is through the macro `_PyErr_IMMUTABLE_WARNING`. This will end up being just one inlined conditional if no handler is set.

The reason to limit the logging is that this particular kind of logging will happen a lot more frequently than the previous ones about instance dict creation.

Reviewed By: tiansihu

Differential Revision: D31775805

fbshipit-source-id: 40b48cd586